+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 ### :bug: Bug Fix
 - `ricos`
   - [#1229](https://github.com/wix-incubator/rich-content/pull/1229) fix(rollup): reduce bundlesize on legacy child support
+- `common`
+  - [#1186](https://github.com/wix-incubator/rich-content/pull/1186) block alignment with indentation
 
 </details>
 <hr/>